\name{calibrate-methods}
\docType{methods}
\alias{calibrate,rrvglm-method}
\alias{calibrate,Coef.qrrvglm-method}
\title{ Calibration for Constrained Regression Models }
\description{
  \code{calibrate} is a generic function applied to QRR-VGLMs and
  RR-VGAMs etc.

}
%\usage{
%  \S4method{calibrate}{cao,Coef.cao}(object, ...)
%}

\section{Methods}{
\describe{

\item{object}{
  The object from which the calibration is performed.

}
}
}
%\note{
%  See \code{\link{lvplot}} which is very much related to biplots.
%
%}

\keyword{methods}
\keyword{classes}
%\keyword{ ~~ other possible keyword(s)}
\keyword{models}
\keyword{regression}



