\name{ccoef-methods}
\docType{methods}
%\alias{ccoef,ANY-method}
\alias{ccoef-method}
\alias{ccoef,cao-method}
\alias{ccoef,Coef.cao-method}
\alias{ccoef,rrvglm-method}
\alias{ccoef,qrrvglm-method}
\alias{ccoef,Coef.rrvglm-method}
\alias{ccoef,Coef.qrrvglm-method}
%
% This does not work:
%\alias{ccoef,cao,Coef.cao,rrvglm,qrrvglm,Coef.rrvglm,Coef.qrrvglm-method}
%
\title{ Constrained (Canonical) Coefficients }
\description{
  \code{ccoef} is a generic function used to return the constrained
  (canonical) coefficients of a  constrained ordination model.
  The function invokes particular methods which depend on the class of
  the first argument.

}
%\usage{
%  \S4method{ccoef}{cao,Coef.cao,rrvglm,qrrvglm,Coef.rrvglm,Coef.qrrvglm}(object, ...)
%}

\section{Methods}{
\describe{

\item{object}{
  The object from which the constrained coefficients are
  extracted.

}
}
}
\keyword{methods}
\keyword{classes}
%\keyword{ ~~ other possible keyword(s)}
\keyword{models}
\keyword{regression}



