%\name{aml}
\name{leukemia}
%\alias{aml}
\alias{leukemia}
\docType{data}
\title{Acute Myelogenous Leukemia Survival Data}
\description{Survival in patients with Acute Myelogenous Leukemia}
\usage{
%data(aml)
data(leukemia)
}
\format{
  \tabular{ll}{
    time:\tab survival or censoring time\cr
    status:\tab censoring status\cr
    x: \tab maintenance chemotherapy given? (factor)\cr
  }
}
\source{
  Rupert G. Miller (1997),
  \emph{Survival Analysis}.
  John Wiley & Sons.
  ISBN: 0-471-25218-2.


}
\note{
  This data set has been transferred from \pkg{survival} and renamed
  from \code{aml} to \code{leukemia}.


}
\keyword{datasets}
