\name{undocumented-methods}
\docType{methods}
%\alias{ccoef,ANY-method}
%\alias{ccoef-method}
%
%
% 202502
\alias{niters,ANY-method}
\alias{niters,vlm-method}
% 202402
\alias{is.zero,NULL-method}
\alias{is.zero,character-method}
\alias{is.zero,logical-method}
% 202401
\alias{vcov,drrvglm-method}
\alias{show,summary.drrvglm-method}
\alias{summary,drrvglm-method}
\alias{coef,summary.drrvglm-method}
\alias{coefficients,summary.drrvglm-method}
% 202311
\alias{show,Coef.drrvglm-method}
\alias{Coef,drrvglm-method}
% 202206
\alias{rqresid,vlm-method}
\alias{rqresid,ANY-method}
\alias{rqresiduals,vlm-method}
\alias{rqresiduals,ANY-method}
% 202111
\alias{KLD,ANY-method}
\alias{KLD,vglm-method}
% 202010
\alias{plotdgaitd,vglm-method}
% 202008
\alias{is.altered,ANY-method}
\alias{is.altered,vglm-method}
\alias{is.inflated,ANY-method}
\alias{is.inflated,vglm-method}
\alias{is.deflated,ANY-method}
\alias{is.deflated,vglm-method}
\alias{is.truncated,ANY-method}
\alias{is.truncated,vglm-method}
\alias{Influence,ANY-method}
\alias{Influence,vgam-method}
\alias{Influence,vglm-method}
\alias{Influence,rrvglm-method}
% 202007
\alias{hdeff,numeric-method}
\alias{hdeff,matrix-method}
% 202004
\alias{showsummaryvglmS4VGAM,ANY,extlogF1-method}
\alias{showvglmS4VGAM,ANY,extlogF1-method}
\alias{showvgamS4VGAM,ANY,extlogF1-method}
\alias{eCDF,vglm-method}
\alias{eCDF,ANY-method}
\alias{fix.crossing,vglm-method}
\alias{fix.crossing,ANY-method}
\alias{is.crossing,vglm-method}
\alias{is.crossing,ANY-method}
% 202003
\alias{altered,ANY-method}
\alias{altered,vglm-method}
\alias{inflated,ANY-method}
\alias{inflated,vglm-method}
\alias{truncated,ANY-method}
\alias{truncated,vglm-method}
\alias{specials,ANY-method}
\alias{specials,vglm-method}
% 202002
\alias{get.offset,ANY-method}
\alias{get.offset,vglm-method}
% 201908
\alias{rootogram4,ANY-method}
\alias{rootogram4,vglm-method}
% 201906
\alias{step4,ANY-method}
\alias{step4,vglm-method}
\alias{add1,vglm-method}
\alias{drop1,vglm-method}
\alias{extractAIC,vglm-method}
\alias{dfterms,ANY-method}
\alias{dfterms,vglm-method}
% 201802
\alias{ordsup,ANY-method}
\alias{ordsup,vglm-method}
\alias{anova,vglm-method}
% 201801
\alias{lrt.stat,ANY-method}
\alias{lrt.stat,vlm-method}
\alias{wald.stat,ANY-method}
\alias{wald.stat,vlm-method}
\alias{score.stat,ANY-method}
\alias{score.stat,vlm-method}
% 20170915
\alias{TIC,ANY-method}
\alias{TIC,vlm-method}
% 201707
% \alias{lrp,vglm-method}
% 201704
\alias{hdeff,vglm-method}
% 201607, 201608:
\alias{psint,pvgam-method}
\alias{summary,pvgam-method}
\alias{show,summary.pvgam-method}
\alias{df.residual,pvgam-method}
\alias{endf,ANY-method}
\alias{endf,pvgam-method}
\alias{endf,summary.pvgam-method}
\alias{vcov,pvgam-method}
\alias{show,pvgam,ANY-method}
\alias{show,pvgam-method}
\alias{model.matrix,pvgam-method}
% 201604:
\alias{plot,pvgam,ANY-method}
% 201602:
% \alias{predictvglmS4VGAM,ANY,binom2.or-method}    20241003 commented this out
% 201601:
\alias{showvglmS4VGAM,ANY,acat-method}
\alias{showvgamS4VGAM,ANY,acat-method}
\alias{showvglmS4VGAM,ANY,multinomial-method}
\alias{showvgamS4VGAM,ANY,multinomial-method}
%
%\alias{coef,vgam-method}
%\alias{coefficients,vgam-method}
% 201512:
% \alias{summaryvglmS4VGAM,ANY,binom2.or-method}    20241003 commented this out
% \alias{showsummaryvglmS4VGAM,ANY,binom2.or-method}    20241003 commented this out
%
\alias{summaryvglmS4VGAM,ANY,posbernoulli.tb-method}
\alias{showsummaryvglmS4VGAM,ANY,posbernoulli.tb-method}
%
\alias{showsummaryvglmS4VGAM,ANY,posbernoulli.b-method}
\alias{showsummaryvglmS4VGAM,ANY,posbernoulli.t-method}
%
\alias{summaryvglmS4VGAM,ANY,VGAMcategorical-method}
\alias{summaryvglmS4VGAM,ANY,cumulative-method}
\alias{summaryvglmS4VGAM,ANY,multinomial-method}
\alias{summaryvglmS4VGAM,ANY,cratio-method}   % 202312
\alias{summaryvglmS4VGAM,ANY,sratio-method}   % 202312
\alias{summaryvglmS4VGAM,ANY,acat-method}   % 202312
%
\alias{showsummaryvglmS4VGAM,ANY,VGAMcategorical-method}
\alias{showsummaryvglmS4VGAM,ANY,cumulative-method}
\alias{showsummaryvglmS4VGAM,ANY,multinomial-method}
\alias{showsummaryvglmS4VGAM,ANY,cratio-method}   % 202312
\alias{showsummaryvglmS4VGAM,ANY,sratio-method}   % 202312
\alias{showsummaryvglmS4VGAM,ANY,acat-method}   % 202312
%
\alias{margeffS4VGAM,ANY,ANY,VGAMcategorical-method}
\alias{margeffS4VGAM,ANY,ANY,VGAMordinal-method}
\alias{margeffS4VGAM,ANY,ANY,acat-method}
\alias{margeffS4VGAM,ANY,ANY,cratio-method}
\alias{margeffS4VGAM,ANY,ANY,sratio-method}
\alias{margeffS4VGAM,ANY,ANY,cumulative-method}
\alias{margeffS4VGAM,ANY,ANY,multinomial-method}
\alias{margeffS4VGAM,ANY,ANY,tobit-method}
% 202209:
\alias{margeffS4VGAM,ANY,ANY,poissonff-method}
\alias{margeffS4VGAM,ANY,ANY,negbinomial-method}
\alias{margeffS4VGAM,ANY,ANY,posnegbinomial-method}
%
%\alias{margeffS4VGAM,ANY,VGAMcategorical-method}
%\alias{margeffS4VGAM,ANY,VGAMordinal-method}
%\alias{margeffS4VGAM,ANY,acat-method}
%\alias{margeffS4VGAM,ANY,cratio-method}
%\alias{margeffS4VGAM,ANY,sratio-method}
%\alias{margeffS4VGAM,ANY,cumulative-method}
%\alias{margeffS4VGAM,ANY,multinomial-method}
%
% 201509:
\alias{term.names,ANY-method}
\alias{term.names,vlm-method}
\alias{responseName,ANY-method}
\alias{responseName,vlm-method}
\alias{has.intercept,ANY-method}
\alias{has.intercept,vlm-method}
% 201508, for R 3.2.2:
\alias{confint,ANY-method}
\alias{confint,vglm-method}
\alias{confint,vgam-method}
\alias{confint,rrvglm-method}
%
% 201503, for R 3.1.3:
\alias{is.buggy,ANY-method}
\alias{is.buggy,vlm-method}
\alias{familyname,ANY-method}
\alias{familyname,vlm-method}
\alias{familyname,vglmff-method}
%
% 201412
\alias{nparam,ANY-method}
\alias{nparam,vlm-method}
\alias{nparam,drrvglm-method}
\alias{nparam,qrrvglm-method}
\alias{nparam,rrvgam-method}
\alias{nparam,vgam-method}
\alias{nparam,vglm-method}
\alias{nparam,rrvglm-method}
\alias{linkfun,ANY-method}
\alias{linkfun,vlm-method}
%
%
% 201407
\alias{concoef,ANY-method}
\alias{concoef,rrvgam-method}
\alias{concoef,Coef.rrvgam-method}
%
%
% 201406
\alias{QR.R,ANY-method}
\alias{QR.R,vglm-method}
\alias{QR.Q,ANY-method}
\alias{QR.Q,vglm-method}
%
%
% 201312
\alias{simulate,ANY-method}
\alias{simulate,vlm-method}
%
% 20131104
\alias{family.name,ANY-method}
\alias{family.name,vlm-method}
\alias{family.name,vglmff-method}
% 20130903
\alias{BIC,ANY-method}
\alias{BIC,vlm-method}
\alias{BIC,vglm-method}
\alias{BIC,vgam-method}
\alias{BIC,rrvglm-method}
\alias{BIC,drrvglm-method}
\alias{BIC,qrrvglm-method}
\alias{BIC,rrvgam-method}
%
% 20121105
\alias{Rank,qrrvglm-method}
\alias{Rank,rrvglm-method}
\alias{Rank,rrvgam-method}
% 20120821
\alias{model.matrix,vsmooth.spline-method}
%
% 20120511
\alias{is.parallel,matrix-method}
\alias{is.parallel,vglm-method}
\alias{is.parallel,ANY-method}
\alias{is.zero,matrix-method}
\alias{is.zero,vglm-method}
\alias{is.zero,ANY-method}
%
%
% 20120215
%\alias{print,vglmff-method}
\alias{show,vglmff-method}
%
%
%
% 20120112
\alias{AIC,ANY-method}
\alias{AICc,ANY-method}
\alias{coef,ANY-method}
\alias{logLik,ANY-method}
\alias{plot,ANY-method}
\alias{vcov,ANY-method}
\alias{plot,rrvgam,ANY-method}
\alias{plot,qrrvglm,ANY-method}
\alias{plot,rcim,ANY-method}
\alias{plot,rcim0,ANY-method}
%\alias{plot,uqo,ANY-method}
\alias{plot,vgam,ANY-method}
\alias{plot,vglm,ANY-method}
\alias{plot,vlm,ANY-method}
\alias{plot,vsmooth.spline,ANY-method}
%
%
%
%
%
\alias{AIC,vlm-method}
\alias{AIC,vglm-method}
\alias{AIC,vgam-method}
\alias{AIC,rrvglm-method}
\alias{AIC,drrvglm-method}
\alias{AIC,qrrvglm-method}
\alias{AIC,rrvgam-method}
\alias{AICc,vlm-method}
% \alias{AICc,vglm-method}  % 20190410
%\alias{AICc,vgam-method}
%\alias{AICc,rrvglm-method}
%\alias{AICc,qrrvglm-method}
\alias{attrassign,lm-method}
\alias{calibrate,ANY-method}
%\alias{calibrate,rrvglm-method}
%\alias{calibrate,qrrvglm-method}
% \alias{calibrate,rrvgam-method}
%\alias{calibrate,uqo-method}
\alias{cdf,vglm-method}
\alias{cdf,vgam-method}
\alias{coefficients,rrvgam-method}
\alias{coefficients,vlm-method}
\alias{coefficients,vglm-method}
\alias{coefficients,qrrvglm-method}
%\alias{coefficients,uqo-method}
\alias{coefficients,vsmooth.spline-method}
\alias{coefficients,vsmooth.spline.fit-method}
\alias{coefficients,summary.vglm-method}
\alias{coefficients,summary.rrvglm-method}
\alias{Coefficients,vlm-method}
\alias{coef,rrvgam-method}
\alias{coef,vlm-method}
\alias{coef,vglm-method}
\alias{coef,qrrvglm-method}
%\alias{coef,uqo-method}
\alias{coef,vsmooth.spline-method}
\alias{coef,vsmooth.spline.fit-method}
\alias{coef,summary.vglm-method}
\alias{coef,summary.rrvglm-method}
\alias{Coef,rrvgam-method}
\alias{Coef,vlm-method}
\alias{Coef,qrrvglm-method}
\alias{Coef,rrvglm-method}
%\alias{Coef,uqo-method}
\alias{constraints,vlm-method}
\alias{deplot,vglm-method}
\alias{deplot,vgam-method}
%
\alias{depvar,ANY-method}
\alias{depvar,rrvgam-method}
\alias{depvar,qrrvglm-method}
\alias{depvar,rcim-method}
\alias{depvar,rrvglm-method}
\alias{depvar,vlm-method}
\alias{depvar,vsmooth.spline-method}
%
\alias{deviance,rrvgam-method}
\alias{deviance,qrrvglm-method}
\alias{deviance,vlm-method}
%\alias{deviance,vglm-method}
%\alias{deviance,uqo-method}
\alias{df.residual,vlm-method}
\alias{effects,vlm-method}
\alias{fitted.values,qrrvglm-method}
\alias{fitted.values,vlm-method}
\alias{fitted.values,vglm-method}
%\alias{fitted.values,uqo-method}
\alias{fitted.values,vsmooth.spline-method}
\alias{fitted,qrrvglm-method}
\alias{fitted,vlm-method}
\alias{fitted,vglm-method}
%\alias{fitted,uqo-method}
\alias{fitted,vsmooth.spline-method}
%
%
%\alias{case.names,ANY-method}
\alias{case.names,vlm-method}
\alias{case.names,vgam-method}
\alias{case.names,vglm-method}
\alias{case.names,rrvglm-method}
\alias{case.names,qrrvglm-method}
\alias{case.names,grc-method}
%
%\alias{variable.names,ANY-method}
\alias{variable.names,vlm-method}
\alias{variable.names,vgam-method}
\alias{variable.names,vglm-method}
\alias{variable.names,rrvglm-method}
\alias{variable.names,qrrvglm-method}
\alias{variable.names,grc-method}
%
%
%\alias{formula,ANY-method}
\alias{formula,vlm-method}
\alias{formula,vgam-method}
\alias{formula,vglm-method}
\alias{formula,rrvglm-method}
\alias{formula,qrrvglm-method}
\alias{formula,grc-method}
%\alias{formula,uqo-method}
% \alias{formula,vsmooth.spline-method}
%
%
%
\alias{hatvalues,ANY-method}
\alias{hatvalues,vlm-method}
\alias{hatvalues,vglm-method}
\alias{hatvalues,rrvgam-method}
\alias{hatvalues,qrrvglm-method}
\alias{hatvalues,rcim-method}
\alias{hatvalues,rrvglm-method}
%
%
\alias{hatplot,ANY-method}
\alias{hatplot,matrix-method}
\alias{hatplot,vlm-method}
\alias{hatplot,vglm-method}
\alias{hatplot,rrvgam-method}
\alias{hatplot,qrrvglm-method}
\alias{hatplot,rcim-method}
\alias{hatplot,rrvglm-method}
%
%
\alias{dfbeta,ANY-method}
\alias{dfbeta,matrix-method}
\alias{dfbeta,vlm-method}
\alias{dfbeta,vglm-method}
\alias{dfbeta,rrvgam-method}
\alias{dfbeta,qrrvglm-method}
\alias{dfbeta,rcim-method}
\alias{dfbeta,rrvglm-method}
%
%
%
\alias{guplot,numeric-method}
\alias{guplot,vlm-method}
%\alias{model.frame,ANY-method}
\alias{model.frame,vlm-method}
%\alias{plot,rcim0,ANY-method}
%\alias{plot,rcim,ANY-method}
%\alias{plot,rrvgam,ANY-method}
%\alias{plot,vlm,ANY-method}
%\alias{plot,vglm,ANY-method}
%\alias{plot,vgam,ANY-method}
%\alias{plot,qrrvglm,ANY-method}
%\alias{plot,uqo,ANY-method}
%\alias{plot,vsmooth.spline,ANY-method}
\alias{predictors,vglm-method}
\alias{rlplot,vglm-method}
\alias{terms,vlm-method}
%\alias{is.bell,uqo-method}
\alias{is.bell,qrrvglm-method}
\alias{is.bell,rrvglm-method}
\alias{is.bell,vlm-method}
\alias{is.bell,rrvgam-method}
\alias{is.bell,Coef.qrrvglm-method}
\alias{logLik,vlm-method}
\alias{logLik,summary.vglm-method}
\alias{logLik,vglm-method}
\alias{logLik,vgam-method}
\alias{logLik,qrrvglm-method}
\alias{logLik,rrvgam-method}
%
\alias{lvplot,rrvgam-method}
\alias{lvplot,qrrvglm-method}
\alias{lvplot,rrvglm-method}
%\alias{lvplot,uqo-method}
%
\alias{lv,rrvglm-method}
\alias{lv,qrrvglm-method}
\alias{lv,rrvgam-method}
\alias{lv,Coef.rrvglm-method}
\alias{lv,Coef.qrrvglm-method}
\alias{lv,Coef.rrvgam-method}
% \alias{lv,uqo-method} defunct
%\alias{latvar,uqo-method}
\alias{latvar,rrvgam-method}
\alias{latvar,Coef.qrrvglm-method}
\alias{latvar,Coef.rrvglm-method}
\alias{latvar,rrvglm-method}
\alias{latvar,qrrvglm-method}
%
\alias{Max,qrrvglm-method}
\alias{Max,Coef.qrrvglm-method}
%\alias{Max,uqo-method}
\alias{Max,rrvgam-method}
\alias{meplot,numeric-method}
\alias{meplot,vlm-method}
%\alias{model.matrix,ANY-method}
\alias{model.matrix,qrrvglm-method}
\alias{model.matrix,vlm-method}
\alias{model.matrix,vgam-method}
\alias{nobs,ANY-method}
\alias{nobs,vlm-method}
\alias{npred,ANY-method}
\alias{npred,vlm-method}
\alias{npred,rrvgam-method}
\alias{npred,qrrvglm-method}
\alias{npred,rcim-method}
\alias{npred,rrvglm-method}
\alias{nvar,ANY-method}
\alias{nvar,vlm-method}
\alias{nvar,vgam-method}
\alias{nvar,rrvglm-method}
\alias{nvar,qrrvglm-method}
\alias{nvar,rrvgam-method}
\alias{nvar,vlm-method}
\alias{nvar,rcim-method}
\alias{Opt,qrrvglm-method}
\alias{Opt,Coef.qrrvglm-method}
%\alias{Opt,uqo-method}
\alias{Opt,rrvgam-method}
\alias{persp,rrvgam-method}
\alias{persp,qrrvglm-method}
%\alias{persp,uqo-method}
\alias{predict,rrvgam-method}
\alias{predict,qrrvglm-method}
\alias{predict,vgam-method}
\alias{predict,vglm-method}
\alias{predict,rrvglm-method}
\alias{predict,vlm-method}
%\alias{predict,uqo-method}
\alias{predict,vsmooth.spline-method}
\alias{predict,vsmooth.spline.fit-method}
%
%
% Added 20090505:
%\alias{print,ANY-method}
%
%
% Added 20111224:
\alias{lrtest,ANY-method}
\alias{lrtest,vglm-method}
%\alias{waldtest,ANY-method}
\alias{print,VGAManova-method}
\alias{show,VGAManova-method}
%
%
%
\alias{print,Coef.rrvgam-method}
\alias{print,summary.rrvgam-method}
\alias{print,qrrvglm-method}
\alias{print,Coef.qrrvglm-method}
\alias{print,rrvglm-method}
% 20090505
\alias{print,summary.qrrvglm-method}
\alias{print,Coef.rrvglm-method}
\alias{print,vlm-method}
\alias{print,vglm-method}
\alias{print,vgam-method}
\alias{print,summary.rrvglm-method}
\alias{print,summary.vgam-method}
\alias{print,summary.vglm-method}
\alias{print,summary.vlm-method}
%\alias{print,uqo-method}
%\alias{print,Coef.uqo-method}
%\alias{print,summary.uqo-method}
\alias{print,vsmooth.spline-method}
\alias{print,rrvgam-method}
\alias{qtplot,vglm-method}
\alias{qtplot,vgam-method}
\alias{residuals,qrrvglm-method}
\alias{residuals,vlm-method}
\alias{residuals,vglm-method}
\alias{residuals,vgam-method}
%\alias{residuals,uqo-method}
\alias{residuals,vsmooth.spline-method}
\alias{resid,qrrvglm-method}
\alias{resid,vlm-method}
\alias{resid,vglm-method}
\alias{resid,vgam-method}
%\alias{resid,uqo-method}
\alias{resid,vsmooth.spline-method}
\alias{show,Coef.rrvgam-method}
\alias{show,summary.rrvgam-method}
\alias{show,qrrvglm-method}
\alias{show,Coef.qrrvglm-method}
\alias{show,rrvglm-method}
% 20090505
\alias{show,summary.qrrvglm-method}
\alias{show,Coef.rrvglm-method}
\alias{show,vlm-method}
\alias{show,vglm-method}
\alias{show,vgam-method}
\alias{show,summary.rrvglm-method}
\alias{show,summary.vgam-method}
\alias{show,summary.vglm-method}
\alias{show,summary.vlm-method}
%\alias{show,uqo-method}
%\alias{show,Coef.uqo-method}
%\alias{show,summary.uqo-method}
\alias{show,vsmooth.spline-method}
\alias{show,rrvgam-method}
\alias{summary,grc-method}
\alias{summary,rrvgam-method}
\alias{summary,qrrvglm-method}
\alias{summary,rcim-method}
\alias{summary,rcim0-method}
\alias{summary,rrvglm-method}
\alias{summary,vgam-method}
\alias{summary,vglm-method}
\alias{summary,vlm-method}
%\alias{summary,uqo-method}
\alias{Tol,rrvgam-method}
\alias{Tol,qrrvglm-method}
\alias{Tol,Coef.qrrvglm-method}
%\alias{Tol,uqo-method}
%\alias{Tol,Coef.uqo-method}
\alias{trplot,qrrvglm-method}
%\alias{trplot,uqo-method}
\alias{trplot,rrvgam-method}
\alias{vcov,rrvglm-method}
\alias{vcov,qrrvglm-method}
\alias{vcov,vlm-method}
\alias{vcov,vglm-method}
\alias{vplot,factor-method}
\alias{vplot,list-method}
\alias{vplot,matrix-method}
\alias{vplot,numeric-method}
\alias{weights,vlm-method}
\alias{weights,vglm-method}
%
%
% This does not work (need one line for each one):
%\alias{trplot,qrrvglm,uqo-method}
%
%
%
\title{ Undocumented Methods Functions }
\description{
  Lots of undocumented methods functions are aliased here.
  In the \pkg{VGAM} package there are currently many
  objects/methods/classes which are currently internal and/or
  undocumented. The help file suppresses the warnings when the package is
  'CHECK'ed.

}
%\usage{
%  \S4method{ccoef}{rrvgam,Coef.rrvgam,rrvglm,qrrvglm,
%                   Coef.rrvglm,Coef.qrrvglm}(object, ...)
%}

\section{Methods}{
  There are many methods and these will be documented over time.

\describe{

\item{object}{
  This argument is often used, and it is the primary object from which the
  function operates on.

}
}
}
\keyword{methods}
\keyword{classes}
%\keyword{ ~~ other possible keyword(s)}
\keyword{models}
\keyword{regression}
\keyword{internal}



