\name{notdocumentedyet}
\alias{notdocumentedyet}
%
%
% 20120310
%\alias{hatvalues}
%\alias{hatvalues.vlm}
%
%
% 20120307
\alias{npred}
\alias{npred.vlm}
%
%
%
% 20120215
% \alias{print.vglmff}
\alias{show.vglmff}
% \alias{print.vfamily}
% \alias{show.Coef.rrar}
% \alias{family.vglm}
\alias{show.vgam}
\alias{show.vglm}
\alias{show.vlm}
% \alias{print.vgam}
% \alias{print.vglm}
% \alias{print.vlm}
% \alias{print.vlm.wfit}
%
%
%
%
% 20120112
\alias{AIC}
\alias{coef}
\alias{logLik}
\alias{plot}
\alias{show.summary.vglm}
\alias{vcov}
\alias{vcovvlm}
\alias{VGAMenv}
\alias{nobs}
\alias{show.Coef.cao}
\alias{show.Coef.qrrvglm}
\alias{show.Coef.rrvglm}
\alias{show.rrvglm}
\alias{show.summary.cao}
% \alias{show.summary.lms}
\alias{show.summary.qrrvglm}
% \alias{show.summary.rc.exponential}
\alias{show.summary.rrvglm}
\alias{show.summary.uqo}
\alias{show.summary.vgam}
\alias{show.summary.vglm}
\alias{show.summary.vlm}
\alias{show.uqo}
\alias{show.vanova}
\alias{show.vsmooth.spline}
%
%
%
%
%
%
%
%
%
% 20111224; lrtest and waldtest stuff
%\alias{lrtest}
%\alias{lrtest_vglm}
%\alias{print_anova}
\alias{update_default}
\alias{update_formula}
%
%\alias{waldtest}
%\alias{waldtest_vglm}
%\alias{waldtest_default}
%\alias{waldtest_formula}
%
%
%
%
% 20110202; 20110317; James Lauder work
%\alias{dexpgeom}
%\alias{pexpgeom}
%\alias{qexpgeom}
%\alias{rexpgeom}
%\alias{expgeometric}
%
%\alias{dweibull3}
%\alias{pweibull3}
%\alias{qweibull3}
%\alias{rweibull3}
%\alias{weibull3}
%
%
%
% 20110321; misc. datasets.
\alias{fibre1.5}
\alias{fibre15}
%
%
% 20120206; for RR-NB, or rrn.tex.
\alias{plota21}
%
%
% 20110202; for Melbourne; these include datasets.
\alias{azprocedure}
\alias{confint_rrnb}
\alias{confint_nb1}
\alias{gala}
\alias{mmt}
%
%
%
%20111128; basics
\alias{is.empty.list}
%
%
%
%
%20101222; Alfian work
%\alias{Rcam}        % Has been written
%\alias{plotrcam0}   % Has been written
%\alias{moffset}     % Has been written
% \alias{Qvar}
\alias{plotqvar}
\alias{depvar.vlm}
%
%
%
%
%20110411
\alias{dbinorm}
%
%20090330
\alias{dclogloglap}
\alias{dlogitlap}
\alias{dprobitlap}
\alias{logitlaplace1.control}
\alias{loglaplace1.control}
\alias{pclogloglap}
\alias{plogitlap}
\alias{pprobitlap}
\alias{qclogloglap}
\alias{qlogitlap}
\alias{qprobitlap}
\alias{rclogloglap}
\alias{rlogitlap}
\alias{rprobitlap}
%
%
%
\alias{A1A2A3}
\alias{AAaa.nohw}
%\alias{AIC}
%\alias{AIC.qrrvglm}
%\alias{AIC.rrvglm}
%\alias{AIC.vgam}
%\alias{AIC.vglm}
%\alias{AIC.vlm}
% \alias{Build.terms}
\alias{Build.terms.vlm}
\alias{Coef.cao}
\alias{Coefficients}
\alias{Cut}
\alias{Deviance.categorical.data.vgam}
\alias{InverseBrat}
\alias{Max.Coef.qrrvglm}
\alias{Max.qrrvglm}
\alias{Opt.Coef.qrrvglm}
\alias{Opt.qrrvglm}
% \alias{R170.or.later}
\alias{Tol.Coef.qrrvglm}
\alias{Tol.Coef.uqo}
\alias{Tol.qrrvglm}
\alias{Tol.uqo}
\alias{a2m}
\alias{abbott}
% \alias{acat.deriv}
% \alias{add.arg}
% \alias{add.constraints}
% \alias{add.hookey}
\alias{add1}
\alias{add1.vgam}
\alias{add1.vglm}
% \alias{adjust.Dmat.expression}
\alias{alaplace1.control}
\alias{alaplace2.control}
\alias{alaplace3.control}
% \alias{alias.vgam}
% \alias{alias.vglm}
\alias{anova.vgam}
\alias{anova.vglm}
% \alias{as.vanova}
% \alias{attrassign}
% \alias{attrassigndefault}
% \alias{attrassignlm}
\alias{beta4}
% \alias{betaffqn}
\alias{biplot}
\alias{biplot.qrrvglm}
% \alias{block.diag}
% \alias{borel.tanner}
\alias{bs}
% \alias{callcaof}
% \alias{callcqof}
% \alias{calldcaof}
% \alias{calldcqof}
% \alias{callduqof}
% \alias{calluqof}
% \alias{canonical.Blist}
% \alias{cao.fit}
\alias{car.all}
% \alias{care.exp}
\alias{ccoef.Coef.cao}
\alias{ccoef.Coef.qrrvglm}
\alias{ccoef.cao}
\alias{ccoef.qrrvglm}
\alias{cdf}
\alias{cdf.lms.bcg}
\alias{cdf.lms.bcn}
\alias{cdf.lms.yjn}
\alias{cdf.vglm}
% \alias{cm.nointercept.vgam}
% \alias{cm.vgam}
% \alias{cm.zero.vgam}
\alias{coefficients}
\alias{coefqrrvglm}
\alias{coefvlm}
\alias{coefvsmooth.spline}
\alias{coefvsmooth.spline.fit}
\alias{constraints.vlm}
% \alias{cqo.fit}
\alias{d2theta.deta2}
% \alias{dcda.fast}
% \alias{dctda.fast.only}
\alias{deplot}
\alias{deplot.default}
\alias{deplot.lms.bcg}
\alias{deplot.lms.bcn}
\alias{deplot.lms.yjn}
\alias{deplot.lms.yjn2}
\alias{deplot.vglm}
\alias{deviance}
\alias{deviance.uqo}
\alias{deviance.vglm}
\alias{deviance.vlm}
%\alias{df.residual}
%\alias{df.residual_vlm}
\alias{dimm}
% \alias{dneg.binomial}
\alias{dnorm2}
\alias{dotC}
\alias{dotFortran}
% \alias{dpsi.dlambda.yjn}
% \alias{drop1.vgam}
% \alias{drop1.vglm}
\alias{dtheta.deta}
% \alias{dy.dyj}
% \alias{dyj.dy}
\alias{effects}
% \alias{effects.vgam}
% \alias{effects.vlm}
% \alias{ei}
% \alias{eij}
\alias{erfc}
\alias{eta2theta}
%\alias{explink}
% \alias{extract.arg}
%\alias{felix}
%\alias{dfelix}
\alias{fff.control}
\alias{fill2}
\alias{fill3}
\alias{fitted}
\alias{fitted.values}
\alias{fitted.values.uqo}
\alias{fittedvsmooth.spline}
%
\alias{variable.names}
\alias{variable.namesvlm}
\alias{variable.namesrrvglm}
\alias{case.names}
\alias{case.namesvlm}
%
\alias{formula}
\alias{formulavlm}
\alias{formulaNA.VGAM}
\alias{gammaff}
% \alias{get.arg}
% \alias{get.rrvglm.se1}
% \alias{get.rrvglm.se2}
% \alias{getind}
% \alias{gh.weight.yjn.11}
% \alias{gh.weight.yjn.12}
% \alias{gh.weight.yjn.13}
% \alias{glag.weight.yjn.11}
% \alias{glag.weight.yjn.12}
% \alias{glag.weight.yjn.13}
% \alias{gleg.weight.yjn.11}
% \alias{gleg.weight.yjn.12}
% \alias{gleg.weight.yjn.13}
\alias{glm}
% \alias{hypersecant}
% \alias{hypersecant.1}
% \alias{ima}
% \alias{interleave.VGAM}
% \alias{invbinomial}
\alias{inverse.gaussianff}
\alias{is.Numeric}
\alias{is.bell}
\alias{is.bell.cao}
\alias{is.bell.qrrvglm}
\alias{is.bell.rrvglm}
\alias{is.bell.vlm}
\alias{Kayfun.studentt}
% \alias{is.linear.term}
% \alias{jitteruqo}
\alias{lm}
\alias{lm2qrrvlm.model.matrix}
\alias{lm2vlm.model.matrix}
\alias{lms.bcg.control}
\alias{lms.bcn.control}
\alias{lms.yjn.control}
\alias{lmscreg.control}
\alias{logLik.vlm}
% \alias{lv.Coef.cao}  20090505
\alias{lv.Coef.qrrvglm}
\alias{lv.cao}
\alias{lv.qrrvglm}
\alias{lvplot.cao}
\alias{m2adefault}
\alias{m2avglm}
% \alias{matrix.power}
\alias{mbesselI0}
\alias{mix2exp.control}
\alias{mix2normal1.control}
\alias{mix2poisson.control}
\alias{model.matrix.qrrvglm}
% \alias{mux11}
% \alias{mux111}
% \alias{mux15}
% \alias{mux2}
% \alias{mux22}
% \alias{mux5}
% \alias{mux55}
% \alias{mux7}
% \alias{mux9}
% \alias{my.dbinom}
\alias{my1}
\alias{my2}
\alias{namesof}
% \alias{natural.ig}
% \alias{neg.binomial}
% \alias{neg.binomial.k}
% \alias{negbin.ab}
% \alias{new.assign}
\alias{nlminbcontrol}
\alias{nbolf2}
\alias{nobs.vlm}
\alias{nvar}
\alias{nvar.vlm}
\alias{nvar.vgam}
\alias{nvar.rrvglm}
\alias{nvar.qrrvglm}
\alias{nvar.cao}
\alias{nvar.rcam}
\alias{ns}
% \alias{num.deriv.rrr}
\alias{persp}
\alias{persp.cao}
\alias{plot.cao}
\alias{plotpreplotvgam}
\alias{plotvglm}
\alias{plotvlm}
\alias{plotvsmooth.spline}
\alias{pnorm2}
% \alias{poissonqn}
\alias{poly}
\alias{predict}
\alias{predict.cao}
\alias{predict.glm}
\alias{predict.lm}
\alias{predict.mlm}
% \alias{predictqrrvglm}
\alias{predict.rrvglm}
\alias{predict.uqo}
\alias{predict.vgam}
\alias{predict.vlm}
\alias{predictcao}
\alias{predictors}
\alias{predictors.vglm}
\alias{predictvsmooth.spline}
\alias{predictvsmooth.spline.fit}
% \alias{preplotvgam}
\alias{print}
\alias{procVec}
\alias{negzero.expression}
\alias{process.binomial2.data.vgam}
\alias{process.categorical.data.vgam}
% \alias{process.constraints}
% \alias{proj.vgam}
% \alias{proj.vglm}
\alias{put.caption}
% \alias{pweights}
% \alias{qrrvglm.xprod}
\alias{qtplot}
\alias{qtplot.default}
\alias{qtplot.lms.bcg}
\alias{qtplot.lms.bcn}
\alias{explot.lms.bcn}
\alias{qtplot.lms.yjn}
\alias{qtplot.lms.yjn2}
\alias{qtplot.vextremes}
\alias{qtplot.vglm}
\alias{quasiff}
\alias{rainfall}
% \alias{remove.arg}
% \alias{replace.constraints}
\alias{resid}
\alias{residuals}
% \alias{residualsqrrvglm}
% \alias{residualsuqo}
% \alias{residualsvglm}
% \alias{residualsvlm}
% \alias{residvsmooth.spline}
\alias{rlplot}
\alias{rlplot.vextremes}
\alias{rlplot.vglm}
% \alias{rrar.Ak1}
% \alias{rrar.Ci}
% \alias{rrar.Di}
% \alias{rrar.Ht}
% \alias{rrar.Mi}
% \alias{rrar.Mmat}
% \alias{rrar.UU}
% \alias{rrar.Ut}
% \alias{rrar.Wmat}
\alias{rrar.control}
% \alias{rrr.alternating.expression}
% \alias{rrr.deriv.gradient.fast}
% \alias{rrr.deriv.rss}
% \alias{rrr.derivC.rss}
% \alias{rrr.derivative.expression}
% \alias{rrr.end.expression}
% \alias{rrr.init.expression}
% \alias{rrr.normalize}
\alias{rrvglm.control.Gaussian}
% \alias{rrvglm.fit}
\alias{rss.vgam}
\alias{s.vam}
\alias{scale.default}
\alias{simple.exponential}
\alias{simple.poisson}
\alias{size.binomial}
\alias{stdze1}
\alias{stdze2}
% \alias{step.vgam}
% \alias{step.vglm}
% \alias{subconstraints}
\alias{summary.cao}
\alias{summary.grc}
\alias{summary.lms}
\alias{summary.qrrvglm}
\alias{summary.rc.exponential}
\alias{summaryrcam}
\alias{summary.rrvglm}
\alias{summary.uqo}
\alias{summaryvgam}
\alias{summaryvglm}
\alias{summaryvlm}
% \alias{tapplymat1}
\alias{terms.vlm}
\alias{theta2eta}
% \alias{trivial.constraints}
% \alias{update.vgam}
% \alias{update.vglm}
% \alias{uqo.fit}
% \alias{valid.vglmff}
% \alias{valid.vknotl2}
% \alias{valt}
% \alias{valt.1iter}
% \alias{valt.2iter}
% \alias{valt.control}
% \alias{varassign}
% \alias{vbacksub}
% \alias{vchol}
% \alias{vchol.greenstadt}
\alias{vcontrol.expression}
% \alias{vcovdefault}
% \alias{vcovqrrvglm}
\alias{vcovrrvglm}
% \alias{vcovvlm}
% \alias{veigen}
% \alias{vellipse}
% \alias{vforsub}
% \alias{vgam.fit}
% \alias{vgam.match}
% \alias{vgam.nlchisq}
% \alias{vgety}
\alias{vglm.fit}
\alias{vglm.garma.control}
\alias{vglm.multinomial.control}
\alias{vglm.multinomial.deviance.control}
\alias{dmultinomial}
\alias{vglm.vcategorical.control}
% \alias{vindex}
% \alias{vlabel}
\alias{vlm}
\alias{vlm.control}
% \alias{vlm.wfit}
\alias{vnonlinear.control}
\alias{vplot}
\alias{vplot.default}
\alias{vplot.factor}
\alias{vplot.list}
\alias{vplot.matrix}
\alias{vplot.numeric}
\alias{vvplot.factor}
\alias{weights}
\alias{Wr1}
\alias{Wr2}
% \alias{wweighted.mean}
\alias{wweights}
% \alias{yformat}
% \alias{ylim.scale}
%
%
%
\alias{Coef.uqo-class}
\alias{cao-class}
\alias{rcam0-class}
\alias{rcam-class}
\alias{grc-class}
\alias{qrrvglm-class}
\alias{summary.qrrvglm-class}
\alias{summary.rrvglm-class}
\alias{summary.vgam-class}
\alias{summary.vglm-class}
\alias{summary.vlm-class}
%%% 20101216 \alias{summary.rcam-class}
%\alias{summary.rcam-class}
%\alias{summaryrcam-class}
\alias{uqo-class}
\alias{vcov.qrrvglm-class}
\alias{vlm-class}
\alias{vlmsmall-class}
\alias{vsmooth.spline-class}
\alias{vsmooth.spline.fit-class}
\alias{Coef.cao-class}
\alias{summary.cao-class}
%
%
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Undocumented and Internally Used Functions and Classes }
\description{
  Those currently undocumented and internally used functions are aliased
  to this help file.
  Ditto for some classes.

}
%\usage{
%normal1(lmean = "identity", lsd = "loge", zero = NULL)
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%  \item{lmean}{
%  Link function applied to the mean.
%  See \code{\link{Links}} for more choices.
%
%  }
%}
\details{
  In the \pkg{VGAM} package there are currently many
  objects/methods/classes which are currently internal and/or
  undocumented. The help file suppresses the warnings when the package is
  'CHECK'ed.

}
\value{
  Each objects/methods/classes may or may not have its own individual value.
  These will be documented over time.

}
%\references{
%}

\author{ T. W. Yee }
%\note{
%
%}
%\seealso{
%    \code{gaussianff},
%    \code{\link{posnormal1}}.
%}
%\examples{
%}
\keyword{models}
\keyword{regression}
\keyword{internal}







