\name{notdocumentedyet}
\alias{notdocumentedyet}
%
%
%
% 201612
\alias{label.cols.y}
\alias{valid.vknotl2}
% 201611
\alias{profilevglm}
\alias{vpairs.profile}
\alias{vplot.profile}
% 201609
\alias{prob.munb.size.VGAM}
\alias{negbinomial.initialize.yj}
% 201607, 201608
\alias{mroot2}
\alias{psint}
\alias{psintpvgam}
\alias{df.residual_pvgam}
\alias{startstoppvgam}
\alias{summary.pvgam-class}
%%%%% \alias{summarypvgam}
%%%%    % \alias{show.summary.pvgam}
\alias{endf}
\alias{endfpvgam}
\alias{vcov.pvgam}
\alias{vcov.pvgam-class}
\alias{vlabel}
\alias{show.pvgam}
\alias{model.matrixpvgam}
% 201606
\alias{gharmonic}
\alias{gharmonic2}
\alias{bisection.basic}
\alias{Zeta.aux}
\alias{deflat.limit.oizeta}
% 201605
\alias{deflat.limit.oipospois}
% 20160418 (keyword: mgcvvgam)
% \alias{ps}
\alias{get.X.VLM.aug}
\alias{psv2magic}
% \alias{psvglm.fit}
% \alias{psvlm.wfit}
\alias{pvgam-class}
% \alias{PS}
%
\alias{checkwz}
\alias{vforsub}
\alias{vbacksub}
\alias{vchol}
\alias{process.constraints}
\alias{mux22}
\alias{mux111}
%
%
% 201602, 201603, 201604:
\alias{genbetaII.Loglikfun4}
\alias{posNBD.Loglikfun2}
\alias{NBD.Loglikfun2}
\alias{AR1.gammas}
\alias{Init.mu}
\alias{.min.criterion.VGAM}
\alias{predictvglmS4VGAM}
% 201601:
\alias{EIM.NB.speciald}
\alias{EIM.NB.specialp}
\alias{EIM.posNB.speciald}
\alias{EIM.posNB.specialp}
\alias{showvglmS4VGAM}
\alias{showvgamS4VGAM}
%\alias{coefvgam}
%
% 201512:
\alias{margeffS4VGAM}
\alias{showsummaryvglmS4VGAM}
\alias{summaryvglmS4VGAM}
\alias{findFirstMethod}
\alias{cratio.derivs}
\alias{subsetarray3}
\alias{tapplymat1}
% 201509, for a bug in car::linearHypothesis() and car:::Anova():
\alias{as.char.expression}
\alias{coef.vlm}
\alias{vcov.vlm}
\alias{model.matrix.vlm}
%\alias{has.intercept}
%\alias{has.interceptvlm}
%\alias{term.names}
%\alias{term.namesvlm}
\alias{responseName}
\alias{responseNamevlm}
%
% 201503, 201504, 201505, 201508;
%\alias{confintvglm}
\alias{qlms.bcn}
\alias{dlms.bcn}
\alias{dbetaII}
% \alias{AR1.control}
% \alias{param.names}  % 20151105
%\alias{is.buggy}
%\alias{is.buggy.vlm}
%
% 201412;
%\alias{linkfun.vglm}
% 201408;
\alias{dlevy}
\alias{plevy}
\alias{qlevy}
\alias{rlevy}
% 201407; expected.betabin.ab is needed for zibetabinomialff() in YBook.
\alias{grid.search}
\alias{grid.search2}
\alias{grid.search3}
\alias{grid.search4}
\alias{expected.betabin.ab}
% 201406;
% \alias{interleave.VGAM}   DONE 20151204
\alias{interleave.cmat}  % 201506;
\alias{marcumQ}
\alias{QR.Q}
\alias{QR.R}
%
%\alias{sm.bs}
%\alias{sm.ns}
%\alias{sm.poly}
%\alias{sm.scale}
%\alias{sm.scale.default}
%
%
%
%
% 201312;
% \alias{simulate.vlm}
% 201311;             20150316: modified to familyname
\alias{familyname}
\alias{familyname.vlm}
\alias{familyname.vglmff}
% 201309;
\alias{I.col}
\alias{BIC}
\alias{check.omit.constant}
%
% 201308;
%\alias{dbiclaytoncop}
%\alias{rbiclaytoncop}
%\alias{biclaytoncop}
%
% 201307;
%\alias{posnormal.control}
\alias{rec.normal.control}
\alias{rec.exp1.control}
%\alias{kendall.tau}
%\alias{binormalcop}
%\alias{dbinormcop}
%\alias{pbinormcop}
%\alias{rbinormcop}
%\alias{expint, expexpint, expint.E1}
%
% 201302;
% \alias{pgamma.deriv.unscaled}
% \alias{pgamma.deriv}
% \alias{digami}
%
% 201212;
\alias{binom2.rho.ss}
%
% 20121105;
% \alias{posbernoulli.b.control}
\alias{N.hat.posbernoulli}
\alias{Rank}
\alias{Rank.rrvglm}
\alias{Rank.qrrvglm}
\alias{Rank.rrvgam}
% 20121015; delete this later
%\alias{huggins91.old}
%
% 20120912
\alias{arwz2wz}
%
% 20120813 New links (no earg)
\alias{Dtheta.deta}
\alias{D2theta.deta2}
\alias{Eta2theta}
\alias{Theta2eta}
\alias{link2list}
\alias{Namesof}
%
%
%
%
% 20120514, 20120528,
\alias{w.wz.merge}
\alias{w.y.check}
\alias{vweighted.mean.default}
%
% 20120418
\alias{nvar_vlm}
% 20120310
%\alias{hatvalues}
%\alias{hatvalues.vlm}
%
%
% 20120307
\alias{npred}
\alias{npred.vlm}
%
%
%
% 20120215
% \alias{print.vglmff}
\alias{show.vglmff}
% \alias{print.vfamily}
% \alias{show.Coef.rrar}
% \alias{family.vglm}
\alias{show.vgam}
\alias{show.vglm}
\alias{show.vlm}
% \alias{print.vgam}
% \alias{print.vglm}
% \alias{print.vlm}
% \alias{print.vlm.wfit}
%
%
%
%
% 20120112
\alias{AIC}
\alias{AICc}
\alias{coef}
\alias{logLik}
\alias{plot}
%\alias{vcov}      % 20150828
%\alias{vcovvlm}      % 20150828
\alias{VGAMenv}
\alias{nobs}
\alias{show.Coef.rrvgam}
\alias{show.Coef.qrrvglm}
\alias{show.Coef.rrvglm}
\alias{show.rrvglm}
\alias{show.summary.rrvgam}
% \alias{show.summary.lms}
\alias{show.summary.qrrvglm}
% \alias{show.summary.rc.exponential}
\alias{show.summary.rrvglm}
%\alias{show.summary.uqo}
% \alias{show.summary.vgam}
% \alias{show.summary.vglm} % 20150831
\alias{show.summary.vlm}
%\alias{show.uqo}
\alias{show.vanova}
\alias{show.vsmooth.spline}
%
%
%
%
%
%
%
%
%
% 20111224; lrtest and waldtest stuff
%\alias{lrtest}
%\alias{lrtest_vglm}
%\alias{print_anova}
\alias{update_default}
\alias{update_formula}
%
%\alias{waldtest}
%\alias{waldtest_vglm}
%\alias{waldtest_default}
%\alias{waldtest_formula}
%
%
%
%
% 20110202; 20110317; James Lauder work
%\alias{dexpgeom}
%\alias{pexpgeom}
%\alias{qexpgeom}
%\alias{rexpgeom}
%\alias{expgeometric}
%
%\alias{dweibull3}
%\alias{pweibull3}
%\alias{qweibull3}
%\alias{rweibull3}
%\alias{weibull3}
%
%
%
% 20110321; misc. datasets.
%\alias{fibre1.5}
%\alias{fibre15}
%
%
% 20120206; for RR-NB, or rrn.tex.
\alias{plota21}
%
%
% 20110202; for Melbourne; these include datasets.
\alias{azprocedure}
\alias{Confint.rrnb}
\alias{Confint.nb1}
%\alias{gala}
% \alias{melbmaxtemp}
%
%
%
%20111128; basics
\alias{is.empty.list}
%
%
%
%
%20101222; Alfian work
%\alias{Rcim}        % Has been written
%\alias{plotrcim0}   % Has been written
%\alias{moffset}     % Has been written
% \alias{Qvar}
\alias{plotqvar}
\alias{qvplot}
\alias{depvar.vlm}
%
%
%
%
%20110411
%\alias{dbinorm}
\alias{dnorm2}
%
%20090330
\alias{dclogloglap}
\alias{dlogitlap}
\alias{dprobitlap}
\alias{logitlaplace1.control}
\alias{loglaplace1.control}
\alias{pclogloglap}
\alias{plogitlap}
\alias{pprobitlap}
\alias{qclogloglap}
\alias{qlogitlap}
\alias{qprobitlap}
\alias{rclogloglap}
\alias{rlogitlap}
\alias{rprobitlap}
%
%
%
% \alias{A1A2A3.orig}  20140909; same as A1A2A3(hwe = TRUE)
\alias{AAaa.nohw}
%\alias{AIC}
%\alias{AIC.qrrvglm}
%\alias{AIC.rrvglm}
%\alias{AIC.vgam}
%\alias{AIC.vglm}
%\alias{AIC.vlm}
% \alias{Build.terms}
\alias{Build.terms.vlm}
\alias{Coef.rrvgam}
\alias{Coefficients}
\alias{Cut}
\alias{Deviance.categorical.data.vgam}
\alias{InverseBrat}
\alias{Max.Coef.qrrvglm}
\alias{Max.qrrvglm}
\alias{Opt.Coef.qrrvglm}
\alias{Opt.qrrvglm}
% \alias{R170.or.later}
\alias{Tol.Coef.qrrvglm}
%\alias{Tol.Coef.uqo}
\alias{Tol.qrrvglm}
%\alias{Tol.uqo}
\alias{a2m}
% \alias{abbott}  % 20150320; no longer releasing family.quantal.R.
% \alias{acat.deriv}
% \alias{add.arg}
% \alias{add.constraints}
% \alias{add.hookey}
\alias{add1}
\alias{add1.vgam}
\alias{add1.vglm}
% \alias{adjust.Dmat.expression}
\alias{alaplace1.control}
\alias{alaplace2.control}
\alias{alaplace3.control}
% \alias{alias.vgam}
% \alias{alias.vglm}
\alias{anova.vgam}
\alias{anova.vglm}
% \alias{as.vanova}
% \alias{attrassign}
% \alias{attrassigndefault}
% \alias{attrassignlm}
% \alias{beta4}
% \alias{betaffqn}
\alias{biplot}
\alias{biplot.qrrvglm}
% \alias{block.diag}
% \alias{borel.tanner}
% \alias{callcaof}
% \alias{callcqof}
% \alias{calldcaof}
% \alias{calldcqof}
% \alias{callduqof}
% \alias{calluqof}
% \alias{canonical.Hlist}
% \alias{cao.fit}
\alias{car.all}
\alias{care.exp}
%
\alias{concoef.Coef.rrvgam}
\alias{concoef.Coef.qrrvglm}
\alias{concoef.rrvgam}
\alias{concoef.qrrvglm}
%
\alias{cdf}
\alias{cdf.lms.bcg}
\alias{cdf.lms.bcn}
\alias{cdf.lms.yjn}
\alias{cdf.vglm}
\alias{cm.VGAM}
\alias{cm.zero.VGAM}
\alias{cm.nointercept.VGAM}
\alias{coefficients}
\alias{coefqrrvglm}
% \alias{coefvlm}  % 20140124
\alias{coefvsmooth.spline}
\alias{coefvsmooth.spline.fit}
% \alias{constraints.vlm}
% \alias{cqo.fit}
\alias{d2theta.deta2}
% \alias{dcda.fast}
% \alias{dctda.fast.only}
\alias{deplot}
\alias{deplot.default}
\alias{deplot.lms.bcg}
\alias{deplot.lms.bcn}
\alias{deplot.lms.yjn}
\alias{deplot.lms.yjn2}
\alias{deplot.vglm}
\alias{deviance}
%\alias{deviance.uqo}
%\alias{deviance.vglm}
\alias{deviance.vlm}
\alias{deviance.qrrvglm}
%\alias{df.residual}
%\alias{df.residual_vlm}
% \alias{dimm}  % 20151105
% \alias{dneg.binomial}
\alias{dnorm2}
%\alias{dotC}
%\alias{dotFortran}
% \alias{dpsi.dlambda.yjn}
% \alias{drop1.vgam}
% \alias{drop1.vglm}
\alias{dtheta.deta}
% \alias{dy.dyj}
% \alias{dyj.dy}
\alias{effects}
% \alias{effects.vgam}
% \alias{effects.vlm}
% \alias{eifun}
% \alias{eijfun}
\alias{eta2theta}
%\alias{explink}
% \alias{extract.arg}
\alias{fff.control}
\alias{fill2}
\alias{fill3}
\alias{fitted}
\alias{fitted.values}
%\alias{fitted.values.uqo}
\alias{fittedvsmooth.spline}
%
\alias{variable.names}
\alias{variable.namesvlm}
\alias{variable.namesrrvglm}
\alias{case.names}
\alias{case.namesvlm}
%
\alias{formula}
\alias{formulaNA.VGAM}
\alias{gammaff}
% \alias{get.arg}
% \alias{get.rrvglm.se1}
% \alias{get.rrvglm.se2}
% \alias{getind}
% \alias{gh.weight.yjn.11}
% \alias{gh.weight.yjn.12}
% \alias{gh.weight.yjn.13}
% \alias{glag.weight.yjn.11}
% \alias{glag.weight.yjn.12}
% \alias{glag.weight.yjn.13}
% \alias{gleg.weight.yjn.11}
% \alias{gleg.weight.yjn.12}
% \alias{gleg.weight.yjn.13}
\alias{glm}
% \alias{hypersecant}
% \alias{hypersecant01}
% \alias{ima}
% \alias{inv.binomial}
\alias{inverse.gaussianff}
\alias{is.Numeric}
\alias{is.bell}
\alias{is.bell.rrvgam}
\alias{is.bell.qrrvglm}
\alias{is.bell.rrvglm}
\alias{is.bell.vlm}
\alias{Kayfun.studentt}
% \alias{is.linear.term}
% \alias{jitteruqo}
\alias{lm}
\alias{lm2qrrvlm.model.matrix}
\alias{lm2vlm.model.matrix}
\alias{vlm2lm.model.matrix}
\alias{lms.bcg.control}
\alias{lms.bcn.control}
\alias{lms.yjn.control}
\alias{lmscreg.control}
% \alias{logLik.vlm}
\alias{logLik.qrrvglm}
% \alias{lv.Coef.rrvgam} 20090505
\alias{latvar.Coef.qrrvglm}
\alias{latvar.rrvgam}
\alias{latvar.rrvglm}
\alias{latvar.qrrvglm}
\alias{lvplot.rrvgam}
\alias{m2a}
%\alias{m2avglm}
% \alias{matrix.power}
\alias{mbesselI0}
\alias{mix2exp.control}
\alias{mix2normal.control}
\alias{mix2poisson.control}
\alias{model.matrix.qrrvglm}
\alias{model.matrixvgam}
% \alias{mux11}
% \alias{mux15}
% \alias{mux2}
% \alias{mux5}
% \alias{mux55}
% \alias{mux7}
% \alias{mux9}
% \alias{my.dbinom}
\alias{my1}
\alias{my2}
\alias{namesof}
% \alias{natural.ig}
% \alias{neg.binomial}
% \alias{neg.binomial.k}
% \alias{negbin.ab}
% \alias{new.assign}
\alias{nlminbcontrol}
\alias{nbolf2}
\alias{nobs.vlm}
\alias{nvar}
\alias{nvar.vlm}
\alias{nvar.vgam}
\alias{nvar.rrvglm}
\alias{nvar.qrrvglm}
\alias{nvar.rrvgam}
\alias{nvar.rcim}
% \alias{num.deriv.rrr}
\alias{persp}
\alias{persp.rrvgam}
\alias{plot.rrvgam}
\alias{plotpreplotvgam}
%\alias{plotvglm}
\alias{plotvlm}
\alias{plotvsmooth.spline}
% \alias{pnorm2} done 20120910
% \alias{poissonqn}
\alias{predict}
\alias{predict.rrvgam}
\alias{predict.glm}
\alias{predict.lm}
\alias{predict.mlm}
% \alias{predictqrrvglm}
\alias{predict.rrvglm}
%\alias{predict.uqo}
\alias{predict.vgam}
\alias{predict.vlm}
\alias{predictrrvgam}
\alias{predictors}
\alias{predictors.vglm}
\alias{predictvsmooth.spline}
\alias{predictvsmooth.spline.fit}
% \alias{preplotvgam}
\alias{print}
\alias{procVec}
\alias{negzero.expression.VGAM}
\alias{process.binomial2.data.VGAM}
\alias{process.categorical.data.VGAM}
% \alias{proj.vgam}
% \alias{proj.vglm}
\alias{put.caption}
% \alias{pweights}
% \alias{qrrvglm.xprod}
\alias{qtplot}
\alias{qtplot.default}
\alias{qtplot.lms.bcg}
\alias{qtplot.lms.bcn}
\alias{explot.lms.bcn}
\alias{qtplot.lms.yjn}
\alias{qtplot.lms.yjn2}
\alias{qtplot.vextremes}
\alias{qtplot.vglm}
\alias{quasiff}
% \alias{rainfall}
% \alias{remove.arg}
% \alias{replace.constraints}
\alias{resid}
\alias{residuals}
% \alias{residualsqrrvglm}
% \alias{residualsuqo}
% \alias{residualsvglm}
% \alias{residualsvlm}
% \alias{residvsmooth.spline}
\alias{rlplot}
\alias{rlplot.vextremes}
\alias{rlplot.vglm}
% \alias{rrar.Ak1}
% \alias{rrar.Ci}
% \alias{rrar.Di}
% \alias{rrar.Ht}
% \alias{rrar.Mi}
% \alias{rrar.Mmat}
% \alias{rrar.UU}
% \alias{rrar.Ut}
% \alias{rrar.Wmat}
\alias{rrar.control}
% \alias{rrr.alternating.expression}
% \alias{rrr.deriv.gradient.fast}
% \alias{rrr.deriv.rss}
% \alias{rrr.derivC.rss}
% \alias{rrr.derivative.expression}
% \alias{rrr.end.expression}
% \alias{rrr.init.expression}
% \alias{rrr.normalize}
% \alias{rrvglm.control.Gaussian}
% \alias{rrvglm.fit}
\alias{ResSS.vgam}
\alias{s.vam}
\alias{simple.exponential}
\alias{better.exponential}
\alias{simple.poisson}
\alias{size.binomial}
%
%
\alias{sm.min1}
\alias{sm.min2}
\alias{sm.scale1}
\alias{sm.scale2}
%\alias{stdze1}
%\alias{stdze2}
%
%
% \alias{step.vgam}
% \alias{step.vglm}
% \alias{subconstraints}
\alias{summary.rrvgam}
\alias{summary.grc}
\alias{summary.lms}
\alias{summary.qrrvglm}
\alias{summary.rc.exponential}
\alias{summaryrcim}
\alias{summary.rrvglm}
%\alias{summary.uqo}
% \alias{summaryvgam}
%\alias{summaryvglm}  % 20150831
\alias{summaryvlm}
% \alias{tapplymat1}
\alias{terms.vlm}
\alias{termsvlm}
\alias{theta2eta}
% \alias{trivial.constraints}
% \alias{update.vgam}
% \alias{update.vglm}
% \alias{uqo.fit}
% \alias{valid.vglmff}
% \alias{valid.vknotl2}
\alias{valt.control}
% \alias{valt}
% \alias{valt.1iter}
% \alias{valt.2iter}
% \alias{valt.control}
% \alias{varassign}
% \alias{vchol.greenstadt}
\alias{vcontrol.expression}
% \alias{vcovdefault}
% \alias{vcovqrrvglm}
%\alias{vcovrrvglm}   % 20150828
% \alias{vcovvlm}
% \alias{veigen}
% \alias{vellipse}
% \alias{vgam.fit}
% \alias{vgam.match}
% \alias{vgam.nlchisq}
% \alias{vgety}
\alias{vgam.fit}
\alias{vglm.fit}
\alias{vglm.garma.control}
\alias{vglm.multinomial.control}
\alias{vglm.multinomial.deviance.control}
\alias{dmultinomial}
\alias{vglm.VGAMcategorical.control}
% \alias{vindex}
% \alias{vlabel}
\alias{vlm}
\alias{vlm.control}
% \alias{vlm.wfit}
\alias{vnonlinear.control}
\alias{vplot}
\alias{vplot.default}
\alias{vplot.factor}
\alias{vplot.list}
\alias{vplot.matrix}
\alias{vplot.numeric}
\alias{vvplot.factor}
\alias{weights}
\alias{Wr1}
\alias{Wr2}
% \alias{wweighted.mean}
\alias{wweights}
% \alias{yformat}
% \alias{ylim.scale}
%
%
%
%\alias{Coef.uqo-class}
\alias{rrvgam-class}
\alias{rcim0-class}
\alias{rcim-class}
\alias{grc-class}
\alias{qrrvglm-class}
\alias{summary.qrrvglm-class}
\alias{summary.rrvglm-class}
\alias{summary.vgam-class}
\alias{summary.vglm-class}
\alias{summary.vlm-class}
%%% 20101216 \alias{summary.rcim-class}
%\alias{summary.rcim-class}
%\alias{summaryrcim-class}
%\alias{uqo-class}
\alias{vcov.qrrvglm-class}
\alias{vlm-class}
\alias{vlmsmall-class}
\alias{vsmooth.spline-class}
\alias{vsmooth.spline.fit-class}
\alias{Coef.rrvgam-class}
\alias{summary.rrvgam-class}
%
%
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Undocumented and Internally Used Functions and Classes }
\description{
  Those currently undocumented and internally used functions are aliased
  to this help file.
  Ditto for some classes.

}
%\usage{
%uninormal(lmean = "identitylink", lsd = "loge", zero = NULL)
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%  \item{lmean}{
%  Link function applied to the mean.
%  See \code{\link{Links}} for more choices.
%
%  }
%}
\details{
  In the \pkg{VGAM} package there are currently many
  objects/methods/classes which are currently internal and/or
  undocumented. The help file suppresses the warnings when the package is
  'CHECK'ed.

}
\value{
  Each objects/methods/classes may or may not have its own individual value.
  These will be documented over time.


}
%\references{
%}

\author{ T. W. Yee }
%\note{
%
%}
%\seealso{
% \code{gaussianff},
% \code{\link{posnormal}}.
%
%}
%\examples{
%}
\keyword{models}
\keyword{regression}
\keyword{internal}







