\name{undocumented-methods}
\docType{methods}
%\alias{ccoef,ANY-method}
%\alias{ccoef-method}
%
%
%
\alias{attrassign,lm-method}
\alias{calibrate,qrrvglm-method}
\alias{calibrate,cao-method}
\alias{calibrate,uqo-method}
\alias{cdf,vglm-method}
\alias{cdf,vgam-method}
\alias{coefficients,cao-method}
\alias{coefficients,vlm-method}
\alias{coefficients,vglm-method}
\alias{coefficients,qrrvglm-method}
\alias{coefficients,uqo-method}
\alias{coefficients,vsmooth.spline-method}
\alias{coefficients,vsmooth.spline.fit-method}
\alias{Coefficients,vlm-method}
\alias{coef,cao-method}
\alias{coef,vlm-method}
\alias{coef,vglm-method}
\alias{coef,qrrvglm-method}
\alias{coef,uqo-method}
\alias{coef,vsmooth.spline-method}
\alias{coef,vsmooth.spline.fit-method}
\alias{Coef,cao-method}
\alias{Coef,vlm-method}
\alias{Coef,qrrvglm-method}
\alias{Coef,rrvglm-method}
\alias{Coef,uqo-method}
\alias{constraints,vlm-method}
\alias{deplot,vglm-method}
\alias{deplot,vgam-method}
\alias{deviance,qrrvglm-method}
\alias{deviance,vlm-method}
\alias{deviance,vglm-method}
\alias{deviance,uqo-method}
\alias{df.residual,vlm-method}
\alias{effects,vlm-method}
\alias{fitted.values,qrrvglm-method}
\alias{fitted.values,vlm-method}
\alias{fitted.values,vglm-method}
\alias{fitted.values,uqo-method}
\alias{fitted.values,vsmooth.spline-method}
\alias{fitted,qrrvglm-method}
\alias{fitted,vlm-method}
\alias{fitted,vglm-method}
\alias{fitted,uqo-method}
\alias{fitted,vsmooth.spline-method}
\alias{guplot,numeric-method}
\alias{guplot,vlm-method}
\alias{model.frame,vlm-method}
\alias{plot,cao,ANY-method}
\alias{plot,vlm,ANY-method}
\alias{plot,vglm,ANY-method}
\alias{plot,vgam,ANY-method}
\alias{plot,qrrvglm,ANY-method}
\alias{predictors,vglm-method}
\alias{rlplot,vglm-method}
\alias{terms,vlm-method}
\alias{is.bell,uqo-method}
\alias{is.bell,qrrvglm-method}
\alias{is.bell,rrvglm-method}
\alias{is.bell,vlm-method}
\alias{is.bell,cao-method}
\alias{is.bell,Coef.qrrvglm-method}
\alias{logLik,vlm-method}
\alias{logLik,summary.vglm-method}
\alias{lvplot,cao-method}
\alias{lvplot,qrrvglm-method}
\alias{lvplot,rrvglm-method}
\alias{lvplot,uqo-method}
\alias{lv,cao-method}
\alias{lv,Coef.cao-method}
\alias{lv,rrvglm-method}
\alias{lv,qrrvglm-method}
\alias{lv,Coef.rrvglm-method}
\alias{lv,Coef.qrrvglm-method}
\alias{lv,uqo-method}
\alias{Max,qrrvglm-method}
\alias{Max,Coef.qrrvglm-method}
\alias{Max,uqo-method}
\alias{meplot,numeric-method}
\alias{meplot,vlm-method}
\alias{model.matrix,qrrvglm-method}
\alias{model.matrix,vlm-method}
\alias{Opt,qrrvglm-method}
\alias{Opt,Coef.qrrvglm-method}
\alias{Opt,uqo-method}
\alias{persp,cao-method}
\alias{persp,qrrvglm-method}
\alias{persp,uqo-method}
\alias{plot,uqo,ANY-method}
\alias{plot,vsmooth.spline,ANY-method}
\alias{predict,cao-method}
\alias{predict,qrrvglm-method}
\alias{predict,vgam-method}
\alias{predict,vglm-method}
\alias{predict,rrvglm-method}
\alias{predict,vlm-method}
\alias{predict,uqo-method}
\alias{predict,vsmooth.spline-method}
\alias{predict,vsmooth.spline.fit-method}
\alias{print,vglmff-method}
\alias{print,Coef.cao-method}
\alias{print,summary.cao-method}
\alias{print,qrrvglm-method}
\alias{print,Coef.qrrvglm-method}
\alias{print,rrvglm-method}
\alias{print,summary.qrrvglm-method}
\alias{print,Coef.rrvglm-method}
\alias{print,vlm-method}
\alias{print,vglm-method}
\alias{print,vgam-method}
\alias{print,summary.rrvglm-method}
\alias{print,summary.vgam-method}
\alias{print,summary.vglm-method}
\alias{print,summary.vlm-method}
\alias{print,uqo-method}
\alias{print,Coef.uqo-method}
\alias{print,summary.uqo-method}
\alias{print,vsmooth.spline-method}
\alias{print,cao-method}
\alias{qtplot,vglm-method}
\alias{qtplot,vgam-method}
\alias{residuals,qrrvglm-method}
\alias{residuals,vlm-method}
\alias{residuals,vglm-method}
\alias{residuals,vgam-method}
\alias{residuals,uqo-method}
\alias{residuals,vsmooth.spline-method}
\alias{resid,vlm-method}
\alias{resid,vglm-method}
\alias{resid,vgam-method}
\alias{resid,uqo-method}
\alias{resid,vsmooth.spline-method}
\alias{show,vglmff-method}
\alias{show,Coef.cao-method}
\alias{show,summary.cao-method}
\alias{show,qrrvglm-method}
\alias{show,Coef.qrrvglm-method}
\alias{show,rrvglm-method}
\alias{show,summary.qrrvglm-method}
\alias{show,Coef.rrvglm-method}
\alias{show,vlm-method}
\alias{show,vglm-method}
\alias{show,vgam-method}
\alias{show,summary.rrvglm-method}
\alias{show,summary.vgam-method}
\alias{show,summary.vglm-method}
\alias{show,summary.vlm-method}
\alias{show,uqo-method}
\alias{show,Coef.uqo-method}
\alias{show,summary.uqo-method}
\alias{show,vsmooth.spline-method}
\alias{show,cao-method}
\alias{summary,grc-method}
\alias{summary,cao-method}
\alias{summary,qrrvglm-method}
\alias{summary,rrvglm-method}
\alias{summary,vgam-method}
\alias{summary,vglm-method}
\alias{summary,vlm-method}
\alias{summary,uqo-method}
\alias{Tol,cao-method}
\alias{Tol,qrrvglm-method}
\alias{Tol,Coef.qrrvglm-method}
\alias{Tol,uqo-method}
\alias{Tol,Coef.uqo-method}
\alias{trplot,qrrvglm-method}
\alias{trplot,uqo-method}
\alias{vcov,rrvglm-method}
\alias{vcov,qrrvglm-method}
\alias{vcov,vlm-method}
\alias{vcov,vglm-method}
\alias{vplot,factor-method}
\alias{vplot,list-method}
\alias{vplot,matrix-method}
\alias{vplot,numeric-method}
\alias{weights,vlm-method}
\alias{weights,vglm-method}
%
%
% This does not work (need one line for each one):
%\alias{trplot,qrrvglm,uqo-method}
%
%
%
\title{ Undocumented Methods Functions }
\description{
  Lots of undocumented methods functions are aliased here.
  In the \pkg{VGAM} package there are currently many
  objects/methods/classes which are currently internal and/or
  undocumented. The help file suppresses the warnings when the package is
  'CHECK'ed.

}
%\usage{
%  \S4method{ccoef}{cao,Coef.cao,rrvglm,qrrvglm,Coef.rrvglm,Coef.qrrvglm}(object, ...)
%}

\section{Methods}{
  There are many methods and these will be documented over time.

\describe{

\item{object}{
  This argument is often used, and it is the primary object from which the
  function operates on.

}
}
}
\keyword{methods}
\keyword{classes}
%\keyword{ ~~ other possible keyword(s)}
\keyword{models}
\keyword{regression}



